lint & tests
